Restaurant-Project/transferencia/requirements.txt
pandas
mysql-connector-python

